se implemento la grabacion total del dpf y algunas mejoras

+++ lib/modules/Deposits from other entities/list_get_other_account_deposit/data/models/list_get_other_account_deposit_model.dart
@@ -50,7 +50,7 @@ class ListGetOtherAccountDepositModel extends ListGetOtherDccountDepositEntity {
         idcState: json["idcState"],
         state: json["state"],
         codMoney: json["codMoney"] ?? '',
-        monto: json["monto"],
+        monto: json["monto"].toDouble(),
         thirdAccount: json["thirdAccount"] ?? '',
         thirdName: json["thirdName"] ?? '',
         accountFunds: json["accountFunds"] ?? '',
