+++ lib/modules/Deposits from other entities/list_get_other_account_deposit/domain/entities/list_get_other_account_deposit_entity.dart
@@ -17,7 +17,7 @@ class ListGetOtherDccountDepositEntity {
   final int idcState;
   final String state;
   final String codMoney;
-  final int monto;
+  final double monto;
   final String thirdAccount;
   final String thirdName;
   final String accountFunds;
